fix: исправление лайков

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -2,6 +2,7 @@ package ru.yandex.practicum.filmorate.service;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.jdbc.support.rowset.SqlRowSet;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
@@ -49,15 +50,27 @@ public class FilmService {
         User user = userStorage.getUserById(userId);
         log.info("Поставлен лайк фильму: {}", filmStorage.getFilmById(film.getId()));
         likeStorage.addLike(film.getId(), user.getId());
+        int rate = film.getRate();
+        int newRate = rate + 1;
+        film.setRate(newRate);
+
+//        film.setRate(filmStorage.getFilmById(filmId).getRate() + 1);
+        filmStorage.update(film);
     }
 
     public void removeLike(long filmId, long userId) {
         Film film = filmStorage.getFilmById(filmId);
         User user = userStorage.getUserById(userId);
-        if (film.getLikes().contains(user.getId())) {
+        int rate = film.getRate();
+        int newRate = rate - 1;
+        film.setRate(newRate);
+
+//        film.setRate(filmStorage.getFilmById(filmId).getRate() - 1);
+        if (film.getLikes() != null && film.getLikes().contains(user.getId())) {
             log.info("Удалён лайк фильму: {}", film);
             likeStorage.deleteLike(film.getId(), user.getId());
         }
+        filmStorage.update(film);
     }
 
     public List<Film> getMostPopularFilms(long count) {
